test_write_shutdown test into tcp test file
